+++ src/main/java/net/minecraft/nbt/NBTTagString.java
@@ -32,7 +32,7 @@ public class NBTTagString extends NBTBase {
     void read(DataInput input, int depth, NBTSizeTracker sizeTracker) throws IOException {
         sizeTracker.read(288L);
         this.data = input.readUTF();
-        sizeTracker.read((long) (16 * this.data.length()));
+        sizeTracker.read(16L * this.data.length());
     }
 
     /**
