+++ src/main/java/net/minecraft/village/VillageDoorInfo.java
@@ -37,7 +37,7 @@ public class VillageDoorInfo {
      * Returns the squared distance between this door and the given coordinate.
      */
     public int getDistanceSquared(int x, int y, int z) {
-        return (int) this.doorBlockPos.distanceSq((double) x, (double) y, (double) z);
+        return (int) this.doorBlockPos.distanceSq(x, y, z);
     }
 
     public int getDistanceToDoorBlockSq(BlockPos pos) {
